% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{Report}
\alias{Report}
\title{Spreadsheet report}
\description{
Spreadsheet report

Spreadsheet report
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wb}}{openxlsx.Workbook object}

\item{\code{path}}{the directory where the report is to be generated}

\item{\code{name}}{the name of the report file}

\item{\code{title}}{Human friendly version of the report name}

\item{\code{author}}{Author that wrote the report}

\item{\code{sheets}}{List that hols sheets object that show in the report}

\item{\code{ts}}{Timestamp when the report was created}

\item{\code{description}}{Description of the contents of the report}

\item{\code{config.filename.ts}}{Whether to add the timestamp to the report file name}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ts}}{Timestamp when the report was created}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Report$new()}}
\item \href{#method-sheet.add}{\code{Report$sheet.add()}}
\item \href{#method-save}{\code{Report$save()}}
\item \href{#method-sheet.pane.freeze}{\code{Report$sheet.pane.freeze()}}
\item \href{#method-sheet.col.width}{\code{Report$sheet.col.width()}}
\item \href{#method-sheet.add.style}{\code{Report$sheet.add.style()}}
\item \href{#method-sheet.merge.header}{\code{Report$sheet.merge.header()}}
\item \href{#method-sheet.conditional.formatting}{\code{Report$sheet.conditional.formatting()}}
\item \href{#method-clone}{\code{Report$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$new(name, path, title, description, author, config.filename.ts = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.add"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.add}{}}}
\subsection{Method \code{sheet.add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.add(sheet)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.pane.freeze"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.pane.freeze}{}}}
\subsection{Method \code{sheet.pane.freeze()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.pane.freeze(name, row, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.col.width"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.col.width}{}}}
\subsection{Method \code{sheet.col.width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.col.width(name, cols, width)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.add.style"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.add.style}{}}}
\subsection{Method \code{sheet.add.style()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.add.style(name, style, rows, cols, expand, stack)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.merge.header"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.merge.header}{}}}
\subsection{Method \code{sheet.merge.header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.merge.header(name, col, row)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sheet.conditional.formatting"></a>}}
\if{latex}{\out{\hypertarget{method-sheet.conditional.formatting}{}}}
\subsection{Method \code{sheet.conditional.formatting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$sheet.conditional.formatting(name, rows, cols, rule, style)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Report$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
