% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.r
\name{Sheet}
\alias{Sheet}
\title{Spreadsheet sheet object}
\description{
Spreadsheet sheet object

Spreadsheet sheet object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the sheet}

\item{\code{title}}{Title of the sheet}

\item{\code{data}}{Data to write into the sheet}

\item{\code{sheet}}{The actual spreadsheet object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sheet$new()}}
\item \href{#method-build.row.spacer}{\code{Sheet$build.row.spacer()}}
\item \href{#method-build.spacer}{\code{Sheet$build.spacer()}}
\item \href{#method-add.format.op}{\code{Sheet$add.format.op()}}
\item \href{#method-pop.char}{\code{Sheet$pop.char()}}
\item \href{#method-build.format.header}{\code{Sheet$build.format.header()}}
\item \href{#method-build.format.freeze.panes}{\code{Sheet$build.format.freeze.panes()}}
\item \href{#method-build.format.cell}{\code{Sheet$build.format.cell()}}
\item \href{#method-build.rule.cell}{\code{Sheet$build.rule.cell()}}
\item \href{#method-build.format.col.width}{\code{Sheet$build.format.col.width()}}
\item \href{#method-build.coord.cell}{\code{Sheet$build.coord.cell()}}
\item \href{#method-process.cell.coords}{\code{Sheet$process.cell.coords()}}
\item \href{#method-process.cell.static}{\code{Sheet$process.cell.static()}}
\item \href{#method-pop.char.header}{\code{Sheet$pop.char.header()}}
\item \href{#method-pop.char.section}{\code{Sheet$pop.char.section()}}
\item \href{#method-config.merge}{\code{Sheet$config.merge()}}
\item \href{#method-clone}{\code{Sheet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the sheet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$new(name, title)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.row.spacer"></a>}}
\if{latex}{\out{\hypertarget{method-build.row.spacer}{}}}
\subsection{Method \code{build.row.spacer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.row.spacer(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.spacer"></a>}}
\if{latex}{\out{\hypertarget{method-build.spacer}{}}}
\subsection{Method \code{build.spacer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.spacer(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add.format.op"></a>}}
\if{latex}{\out{\hypertarget{method-add.format.op}{}}}
\subsection{Method \code{add.format.op()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$add.format.op(op)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop.char"></a>}}
\if{latex}{\out{\hypertarget{method-pop.char}{}}}
\subsection{Method \code{pop.char()}}{
Creates a population characteristics table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$pop.char(
  sections,
  groups,
  header,
  config = NULL,
  styles = NULL,
  rules = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.format.header"></a>}}
\if{latex}{\out{\hypertarget{method-build.format.header}{}}}
\subsection{Method \code{build.format.header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.format.header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.format.freeze.panes"></a>}}
\if{latex}{\out{\hypertarget{method-build.format.freeze.panes}{}}}
\subsection{Method \code{build.format.freeze.panes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.format.freeze.panes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.format.cell"></a>}}
\if{latex}{\out{\hypertarget{method-build.format.cell}{}}}
\subsection{Method \code{build.format.cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.format.cell(row, col, styles)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.rule.cell"></a>}}
\if{latex}{\out{\hypertarget{method-build.rule.cell}{}}}
\subsection{Method \code{build.rule.cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.rule.cell(row, col, rules)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.format.col.width"></a>}}
\if{latex}{\out{\hypertarget{method-build.format.col.width}{}}}
\subsection{Method \code{build.format.col.width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.format.col.width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build.coord.cell"></a>}}
\if{latex}{\out{\hypertarget{method-build.coord.cell}{}}}
\subsection{Method \code{build.coord.cell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$build.coord.cell(s = NULL, g = NULL, a = NULL, v = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process.cell.coords"></a>}}
\if{latex}{\out{\hypertarget{method-process.cell.coords}{}}}
\subsection{Method \code{process.cell.coords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$process.cell.coords(row, col, s = NULL, g = NULL, a = NULL, v = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process.cell.static"></a>}}
\if{latex}{\out{\hypertarget{method-process.cell.static}{}}}
\subsection{Method \code{process.cell.static()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$process.cell.static(row, col, coord)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop.char.header"></a>}}
\if{latex}{\out{\hypertarget{method-pop.char.header}{}}}
\subsection{Method \code{pop.char.header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$pop.char.header(row, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop.char.section"></a>}}
\if{latex}{\out{\hypertarget{method-pop.char.section}{}}}
\subsection{Method \code{pop.char.section()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$pop.char.section(row, col, section)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config.merge"></a>}}
\if{latex}{\out{\hypertarget{method-config.merge}{}}}
\subsection{Method \code{config.merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$config.merge(extra, base)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sheet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
