% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.r
\name{Stat}
\alias{Stat}
\title{Handy Statistics}
\description{
Handy Statistics

Handy Statistics
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Stat$new()}}
\item \href{#method-preety}{\code{Stat$preety()}}
\item \href{#method-avg}{\code{Stat$avg()}}
\item \href{#method-std}{\code{Stat$std()}}
\item \href{#method-sum}{\code{Stat$sum()}}
\item \href{#method-nopp}{\code{Stat$nopp()}}
\item \href{#method-blank}{\code{Stat$blank()}}
\item \href{#method-id}{\code{Stat$id()}}
\item \href{#method-p50}{\code{Stat$p50()}}
\item \href{#method-iqr}{\code{Stat$iqr()}}
\item \href{#method-prev}{\code{Stat$prev()}}
\item \href{#method-count}{\code{Stat$count()}}
\item \href{#method-clone}{\code{Stat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$new(digits = 2, nsmall = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preety"></a>}}
\if{latex}{\out{\hypertarget{method-preety}{}}}
\subsection{Method \code{preety()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$preety(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-avg"></a>}}
\if{latex}{\out{\hypertarget{method-avg}{}}}
\subsection{Method \code{avg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$avg(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-std"></a>}}
\if{latex}{\out{\hypertarget{method-std}{}}}
\subsection{Method \code{std()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$std(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sum"></a>}}
\if{latex}{\out{\hypertarget{method-sum}{}}}
\subsection{Method \code{sum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$sum(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nopp"></a>}}
\if{latex}{\out{\hypertarget{method-nopp}{}}}
\subsection{Method \code{nopp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$nopp(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blank"></a>}}
\if{latex}{\out{\hypertarget{method-blank}{}}}
\subsection{Method \code{blank()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$blank(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-id"></a>}}
\if{latex}{\out{\hypertarget{method-id}{}}}
\subsection{Method \code{id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$id(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p50"></a>}}
\if{latex}{\out{\hypertarget{method-p50}{}}}
\subsection{Method \code{p50()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$p50(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr"></a>}}
\if{latex}{\out{\hypertarget{method-iqr}{}}}
\subsection{Method \code{iqr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$iqr(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prev"></a>}}
\if{latex}{\out{\hypertarget{method-prev}{}}}
\subsection{Method \code{prev()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$prev(n, t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count"></a>}}
\if{latex}{\out{\hypertarget{method-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$count(df, col)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
